+++ src/main/java/com/gamecrew/gamecrew_project/domain/user/dto/request/EmailCodeRequestDto.java
@@ -1,0 +1,9 @@
+package com.gamecrew.gamecrew_project.domain.user.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class EmailCodeRequestDto {
+    private String code;
+    private String email;
+}
