+++ src/main/java/com/gamecrew/gamecrew_project/global/response/constant/Message.java
@@ -2,7 +2,7 @@ package com.gamecrew.gamecrew_project.global.response.constant;
 
 public class Message {
     public static final String SIGNUP_SUCCESSFUL = "회원가입을 완료하였습니다.";
-    public static final String LOGIN_SUCCESSFUL = "로그인에 성공하였습니다.";
+    public static final String AUTH_SUCCESSFUL = "이메일 인증에 성공하였습니다.";
     public static final String AVAILABLE_NICKNAME = "사용 가능한 닉네임입니다.";
     public static final String AVAILABLE_EMAIL = "사용 가능한 이메일입니다.";
     public static final String POST_SUCCESSFUL = "등록 완료";
